fix: resolve 15 mypy type errors across 4 files
- utils/helpers.py: change format_bytes param from int to float
- providers/provider_manager.py: add GeocodeProvider Protocol and type self._providers dict
- processing/geocoder.py: change stats return type from Dict[str, int] to Dict[str, Any]
- main.py: add Optional[] to 6 function params with None defaults, annotate 2 list vars
- Pin mypy==1.14.1 and add types-requests>=2.32.0 to requirements.txt

+++ utils/helpers.py
@@ -26,7 +26,7 @@ def setup_directories() -> None:
         d.mkdir(parents=True, exist_ok=True)
 
 
-def format_bytes(size: int) -> str:
+def format_bytes(size: float) -> str:
     """Format byte count to human-readable string."""
     for unit in ["B", "KB", "MB", "GB"]:
         if size < 1024:
